requirements.txt
asgiref==3.11.0
Django==5.2.10
django-axes==8.3.1
django-environ==0.12.0
holidays==0.93
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
charset-normalizer==3.4.6
pillow==12.2.0
reportlab==4.4.10
